Update LICENSE
To make clear to anyone that thinks I'm trying to claim "ownership" of this, I've removed my username.  The AI chats say I'm supposed to put my name in this as a contributor for my modifications (anytime you see my username in the code, it was suggested by AI), but it doesn't matter to me.  I just want to learn how to do this right, and I will defer to human criticism.  The original project lists ISC license in the package.json so I'm trying to adhere to that.